First Version
Has just the static data table populate and only makes a select in this table. Counts with the database code to sqLite database, and has a simple python server that responds to a GET request with a POST request doing a SELECT on the static table, and selecting the first line when the code given is 1.